+++ src/js/functions/errMsg.mjs
@@ -1,8 +1,9 @@
 /**
  * handles error messages. input element to hold message. will display and hide messages.  
- * @param {object} element 
- * @param {string} msg 
+ * @param {Object} element this object holds and displays messages
+ * @param {Array} msgArr this is the array of messages that will be displayed
  */
+
 export default function errMsg(element, msgArr) {
   element.innerText = "";
     element.classList.add("d-none");
